Move files and repair @martinholmer's test (merge conflict removed it)
