--- tax_microdata_benchmarking/utils/uprating/cbo_population_forecast.yaml
@@ -1,72 +1,0 @@
-# Sources:
-# CBO Jan 2024 publication "The Demographic Outlook: 2024 to 2054"
-# URL: https://www.cbo.gov/publication/59697
-# (a) Data Underlying Figures, Figure 2 spreadsheet tab,
-#     Total Population (millions) [for years from 2021 through 2054]
-# (b) Demographic Projections, Population and growth spreadsheet tab,
-#     Total Social Security area population (millions) [for years after 2054]
-2010: 313.985
-2011: 316.096
-2012: 318.030
-2013: 319.857
-2014: 321.766
-2015: 324.530
-2016: 327.042
-2017: 328.899
-2018: 330.808
-2019: 332.199
-2020: 333.263
-2021: 334.181
-2022: 335.458
-2023: 338.442
-2024: 342.295
-2025: 346.234
-2026: 349.489
-2027: 351.948
-2028: 353.683
-2029: 355.387
-2030: 357.058
-2031: 358.692
-2032: 360.285
-2033: 361.832
-2034: 363.330
-2035: 364.774
-2036: 366.164
-2037: 367.499
-2038: 368.777
-2039: 369.998
-2040: 371.161
-2041: 372.266
-2042: 373.315
-2043: 374.308
-2044: 375.250
-2045: 376.144
-2046: 376.993
-2047: 377.801
-2048: 378.573
-2049: 379.313
-2050: 380.026
-2051: 380.718
-2052: 381.394
-2053: 382.062
-2054: 382.726
-2055: 383.393
-2056: 384.067
-2057: 384.752
-2058: 385.451
-2059: 386.167
-2060: 386.898
-2061: 387.644
-2062: 388.404
-2063: 389.175
-2064: 389.955
-2065: 390.739
-2066: 391.523
-2067: 392.302
-2068: 393.073
-2069: 393.831
-2070: 394.572
-2071: 395.291
-2072: 395.984
-2073: 396.650
-2074: 397.286
